LayoutTests/inspector/console/console-trim-long-urls-expected.txt

CONSOLE MESSAGE: line 11: The URL is: %s
CONSOLE MESSAGE: line 12: http://example.com/2---------3---------4---------5---------6---------7---------8---------9---------0---------1---------2---------3---------4---------5---------6---------7---------8---------9---------0
Tests that a URL logged to the console is trimmed down to 150 characters.

The URL is: http://example.com/2---------3---------4---------5---------6---------7-----…---3---------4---------5---------6---------7---------8---------9---------0 console-trim-long-urls.html:11
http://example.com/2---------3---------4---------5---------6---------7-----…---3---------4---------5---------6---------7---------8---------9---------0 console-trim-long-urls.html:12
